Add Accept-Encoding header for udf submit
This is a workaround to the TileDB-Cloud-Py autogenerated client not
providing a way to set supported compression on the http request. By
adding a header to the spec this becomes an option for the kwargs of the
Apply function.

urllib3 happens to also check for the `Content-Encoding` response header
and auto decodes the response data