Add schemas and endpoints for generic API creation and management in v2.
This creates the new enum `CloudProvider` and the new type
`AccessCredential` (where one user/password pair is the singular), which
is a union type to represent the credential needed to access an array's
data on any single cloud provider.

Also creates endpoints to perform all the usual CRUD operations on
`AccessCredential`s.