Make AccessCredential provider_default and credential types nullable.
Unless we mark this as nullable, our generated code on the server side
cannot easily discern between a user wanting to set a credential to
not be the default (i.e. `"provider_default": false`) vs. not changing
the default status at all (no `provider_default` field).

We need similar support for our credential types as well.