Add stored_param_uuids to UDF APIs for future support.
This will allow the client to send a list of previous UDF result IDs
to be fetched and provided by the server, so that the client does not
need to re-send all of its previous results back to the server.

Then, support can be added on the UDF environment, server, and client.