Remove result_format_version from all APIs (#265)
`result_format_version` is currently unused; we just pass it among
our various components but never do anything with it. If we did live in
a world with multiple versions for a single result format, an empty
`result_format_version` would not be useful, since unless we define
the empty `result_format_version` as a v0 of sorts, we don’t actually
know what format it is and cannot unambiguously decode it.

Take, for example, the incompatiblity of pickles between py3.7 and 3.8:
it doesn’t matter that they’re both pickles; just that the one cannot
decode the other.

Incompatible result format versions are best represented as different
`result_format` values, since fundamentally we have to handle an
incompatible version the exact same way we would have to handle a
different format entirely. In situations where a later version of a
decoder may be able to decode an earlier version of a format, we can
manually track those mappings (because even `result_format_version`
provides no information about that right now).