Create API endpoints for task graph logging.
This adds messages and endpoints for logging task graph details.
When creating a task graph, clients submit the structure of the graph
to the server so that `ArrayTask`s can be associated with the graph
they belong to.

Querying and retrieval work analogously to the same actions for
`ArrayTask`s.