Add TaskGraphRunLocation to distinguish local nodes.
Currently, the UI can't tell between a task graph node that was executed
client-side versus a node that will be executed server-side but hasn't
yet run, leaving difficult-to-parse "idle" nodes across the graph in
situations where a user executes part of a task graph locally.

This creates an enum that allows us to distinguish server-side nodes
from client-side nodes from (forthcoming) "virtual" nodes that represent
data in the graph that is not actually "executed".