Add `friendly_ranges` to `TGQueryRanges` for user-friendlier API.
Because `ranges` specified exactly the range format required by
a UDF call, it meant that task graph API users would be forced to
provide their requested ranges in the form `[lo1, hi1, lo2, hi2, ...]`.

To allow us to provide a better API, we need to also be able to provide
a way for users to provide ranges in a friendlier format, and because
users may provide some values as Nodes, they can't immediately be
translated to the canonical format at graph-building/-registration time.