Replace `swagger-cli` with `go-swagger`.
This PR replaces the swagger-cli-based Azure pipeline with a
go-swagger-based GitHub workflow.  Ever since Nov 2023, swagger-cli is
archived and the pipeline failed to catch a schema issue that I ran into
while using go-swagger.  We appear to be using go-swagger to generate
Cloud-UI's and Cloud-REST's model files, so we should use the very same
tool to validate the spec files.