Add support for AWS KMS key ID with credentials
This adds a new parameter to AWSCredential/AWSRole to allow a user to
associate a KMS key arn with the credential to be used. This allows the
system to support KMS based s3 encryption in a more automated fashion
without the user having to pass the key arn for each request manually.